0.4

Closes #4, #5, #8 and #9

See merge request sherifnasser/AinLanguage!3

+++ include/expression/value/DoubleValue.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include "PrimitiveValue.hpp"
-class DoubleValue:public PrimitiveValue<long double>{
+class DoubleValue:public PrimitiveValue<double>{
     public:
-        DoubleValue(long double value);
+        DoubleValue(double value);
         std::wstring toString()override;
 };
